Extract logic so command is at the upper level
